requirements.txt
Flask==3.1.3
Pillow==12.1.1
gunicorn==23.0.0
requests==2.32.3
firebase-admin==6.5.0
